Debugging

+++ trie.h
@@ -1,4 +1,4 @@
 int triInsert(char* key, float value);
 int tri_insert(char* key, float value);
-int triSearch(char* word, float value);
+float triSearch(char* word, float value);
 int triRemove(char* key);
